Combine throttle implementations
